Fix "login box login dialog box will not display password field"
Similar to what happened in https://bugs.launchpad.net/ubuntu/+source/unity-greeter/+bug/1415168

When the greeter first starts, the login box is too small and its entry isn't visible.

After Escape is pressed or after selecting another user, the login box behaves correctly.

It's worth noting that this issue happened prior to removing indicator support if no indicators were loaded and nm-applet wasn't started.

This leads to believe that the issue existed in unity-greeter already and is related to a race condition.

This commit doesn't properly fix the issue. What it does is call the same code that is called when the user presses Escape.

Note: This issue is not reproducible in test mode.