background: Don't realize() this immediately - only start the image gathering thread during initialization.  Wait for the main window to be realized before passing its cairo surface to the background object and drawing the background for the first time.
We also need to make sure in our background's size_allocate function
to abort if we're not really active yet (no GdkWindow realized, and
no AnimateTimer yet).

Moved some background init stuff from the main window to the background
class for clarity, and removed the default_background setting - it is
already grabbing the default color in Background's class init.