Fix background if image file is not readable
This fixes a bug, where we fell back to the defined background color, if
reading an image file failed.
If that file was a user background, we didn't fall back to the defined
system background before using the color, now we do.
Only if that one is not readable either, we fall back to the color.