[Greeter] should be on top of the config lines (#221)
Some people like me are definitely going to copy the code and paste it in slick-greeter.conf file. Then they will find themselves to be in a situation where uncommenting lines wouldn't work due to the position of [Greeter].

I only found out this because I checked out the lightdm-settings app. In fact I was going to switch to other greeters but just wanted to try something out again becaue to me slick-greeter is the best out-of-the-box greeter experience ever.

Happy to contribute.