avatar: Remove Cinnamon fallback
This makes the behavior of slick-greeter different based on whether
or not Cinnamon is installed on the OS.

If we're to use a default avatar picture, it should be shipped by
slick-greeter, or even better yet by xapp-common (so that DE tools
can use it as well).

Showing no avatar when none is set also kinda looks nice and gives
more freedom to users.